requirements.txt
requests
bs4
pandas
openai
openpyxl
google_play_scraper
certifi
python-dotenv